refactor: move placement details from 'coords' to their own module
