+++ src/lib/coords.ts
@@ -2,11 +2,6 @@ import { Result } from "./result";
 
 type Coords = [x: number, y: number];
 
-type Placement = {
-  position: Coords;
-  orientation: "h" | "v";
-};
-
 function convertCoords(dimensions: Coords) {
   const [width] = dimensions;
 
@@ -28,16 +23,6 @@ function convertIndex(dimensions: Coords) {
   };
 }
 
-function placeShip(location: Placement, length: number) {
-  const [x, y] = location.position;
-
-  if (location.orientation === "h") {
-    return Array.from({ length }, (_, i) => [x + i, y]);
-  }
-
-  return Array.from({ length }, (_, i) => [x, y + i]);
-}
-
 function parseCoords(text: string): Result<Coords, Error> {
   const parsed = text.split(/(?:\D+)/);
 
@@ -56,5 +41,5 @@ function parseCoords(text: string): Result<Coords, Error> {
   };
 }
 
-export { convertCoords, convertIndex, parseCoords, placeShip };
-export type { Coords, Placement };
+export { convertCoords, convertIndex, parseCoords };
+export type { Coords };
